Move ArrayBoolSearchContext to header
